+++ series_intro_recognizer/services/interval_improver.py
@@ -9,7 +9,7 @@ def _filter_too_long_intervals(interval: Interval, cfg: Config) -> Interval:
     If the interval is too long, it is replaced with NaNs.
     """
     return (interval
-            if interval.end - interval.start <= cfg.MAX_SEGMENT_LENGTH_SEC
+            if interval.end - interval.start <= cfg.max_segment_length_sec
             else Interval(math.nan, math.nan))
 
 
@@ -18,15 +18,15 @@ def _adjust_borders(interval: Interval, audio_duration: float, cfg: Config) -> I
     If the interval is too close to the beginning or the end of the audio,
     it adjusts the interval to the beginning or the end of the audio.
     """
-    if cfg.ADJUSTMENT_THRESHOLD is False:
+    if cfg.adjustment_threshold is False:
         return interval
 
     start = 0 \
-        if interval.start - cfg.ADJUSTMENT_THRESHOLD_SECS <= 0 \
+        if interval.start - cfg.adjustment_threshold_secs <= 0 \
         else interval.start
 
     end = audio_duration \
-        if interval.end + cfg.ADJUSTMENT_THRESHOLD_SECS >= audio_duration \
+        if interval.end + cfg.adjustment_threshold_secs >= audio_duration \
         else interval.end
 
     return Interval(start, end)